Restructured package and updated terminology
