+++ resolve_proxy_encoder/worker/celery.py
